+++ public/services/list-services.js
@@ -1,6 +1,8 @@
 import 'isomorphic-fetch'
+import status from './status'
 
 export function createList (list, user, cb) {
+  if (user.username === 'mrormrstestperson@taskmastr.co') return cb(null, { success: true })
   return window.fetch('/lists/create', {
     method: 'put',
     credentials: 'same-origin',
@@ -10,14 +12,8 @@ export function createList (list, user, cb) {
     },
     body: JSON.stringify({ list, user })
   })
-  .then((response) => {
-    if (response.status !== 200) {
-      let error = new Error(response.statusText)
-      error.response = response
-      throw error
-    }
-    return response.json()
-  })
+  .then(status)
+  .then((response) => response.json())
   .then((response) => {
     cb(null, response)
   })
@@ -35,14 +31,8 @@ export function getList (id, cb) {
       'Content-Type': 'application/json'
     }
   })
-  .then((response) => {
-    if (response.status !== 200) {
-      let error = new Error(response.statusText)
-      error.response = response
-      throw error
-    }
-    return response.json()
-  })
+  .then(status)
+  .then((response) => response.json())
   .then((response) => {
     cb(null, response)
   })
@@ -52,6 +42,7 @@ export function getList (id, cb) {
 }
 
 export function removeList (id, user, cb) {
+  if (user.username === 'mrormrstestperson@taskmastr.co') return cb(null, { success: true })
   return window.fetch(`/lists/${id}/delete`, {
     method: 'delete',
     credentials: 'same-origin',
@@ -61,14 +52,8 @@ export function removeList (id, user, cb) {
     },
     body: JSON.stringify({ user })
   })
-  .then((response) => {
-    if (response.status !== 200) {
-      let error = new Error(response.statusText)
-      error.response = response
-      throw error
-    }
-    return response.json()
-  })
+  .then(status)
+  .then((response) => response.json())
   .then((response) => {
     cb(null, response)
   })
@@ -79,6 +64,7 @@ export function removeList (id, user, cb) {
 }
 
 export function updateList (user, listId, listBody, cb) {
+  if (user.username === 'mrormrstestperson@taskmastr.co') return cb(null, { success: true })
   return window.fetch(`/lists/${listId}/update`, {
     method: 'post',
     credentials: 'same-origin',
@@ -88,14 +74,8 @@ export function updateList (user, listId, listBody, cb) {
     },
     body: JSON.stringify({ user, listId, listBody })
   })
-  .then((response) => {
-    if (response.status !== 200) {
-      let error = new Error(response.statusText)
-      error.response = response
-      throw error
-    }
-    return response.json()
-  })
+  .then(status)
+  .then((response) => response.json())
   .then((response) => {
     cb(null, response)
   })
